# lib/api/openapi.yaml
openapi: 3.0.2
paths:
...
  /cart/{user_id}:
    summary: Manage cart
    description: Perform CRUD operations on cart matching user_id
    get:
      summary: Get cart
      description: Get cart matching user_id
      operationId: get_cart
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: number
      responses:
        '200':
          description: GET cart
          content:
            application/json:
              examples:
                Example:
                  value: |-
                    {
                        "id": 1,
                        "user_id": 1,
                        "items": [
                            {
                                "product_id": 54
                            },
                            {
                                "product_id": 27
                            },
                            {
                                "product_id": 3
                            }
                        ]
                    }
        default:
          description: Default error sample response
      tags:
        - cart
    patch:
      summary: Update cart
      description: Update cart matching id
      operationId: update_cart
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: number
      responses:
        default:
          description: Default error sample response
      tags:
        - cart
...
  /checkout/{user_id}:
    summary: Checkout path
    description: Endpoint used for confirming a purchase once payment has succeeded
    post:
      summary: Checkout
      description: Checkout the user's cart
      operationId: checkout
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: number
      responses:
        default:
          description: Default error sample response
      tags:
        - cart
...
info:
  title: kapi
  version: 1.0.0
  description: An e-commerce REST API using Express, Node.js, and Postgres